requirements.txt
numpy
pandas
scikit-learn
scipy
matplotlib
torch
fair-esm
